feat: criando o contexto pra gerenciar os estados
